Skipjack transaction_id field could be blank if the return_code is something other than 1.
If you submit a transaction that fails because of the way the data is
submitted, or the format of said data, then the transaction_id will be
blank. Transaction_id get's filled, it would appear, when a transaction
is submitted to the Merchant Gateway your account uses once Skipjack is
happy with the details submitted. Of course this does not appear to be
documented particularly clearly.  Also a few documentation updates, and
renaming `created` to `creation_date` on the Response model.