Transaction status change methods now operational.
This required changing the primary key structure on the Transaction
model - Skipjack changes the transaction id when a transaction moves
from Authorized to Settled making usage of that field as the primary key
impossible. We now provide a mechanism to update the transaction_id in
the Transaction object with calls to Transaction.get_status() and
Transaction.update_status().